feat: return domain and SCM platform information
Return information about the fully qualified domain name, such as
`github.com` for `https://github.com/user/project`, and the SCM
platform, such as `github` or `gitlab`.

BREAKING CHANGE:

If user or project information cannot be extracted from the URL, then
an object is returned, instead of `null`, and the object's `user` and
`project` properties are set to `null`.