refactor(index): switch to new URL parser
Switch to Node's new `url.URL` parser for parsing repository urls.

We were using `url.parse`, which is now deprecated in NodeJS.

BREAKING CHANGE:

Switch to Node's new `url.URL` parser, and use `example.com` as a base
domain when repository URLs are passed into the module without a
domain associated with them.

For example, if `user/project` is passed to `@hutson/parse-repository-url`,
we set the _base_ domain for `url.URL` to `example.com`, which causes
the `domain` property on the returned object to be set to `example.com`.